Design(React): CheckBox css 추가 및 코드 스타일 수정
